Split fields into classes, remove duplication
