+++ app/forms/field_factory.py
@@ -1,11 +1,11 @@
-from app.forms.handlers.date_handler import DateHandler
-from app.forms.handlers.dropdown_handler import DropdownHandler
-from app.forms.handlers.duration_handler import DurationHandler
-from app.forms.handlers.number_handler import NumberHandler
-from app.forms.handlers.select_handler import SelectHandler
-from app.forms.handlers.select_multiple_handler import SelectMultipleHandler
-from app.forms.handlers.string_handler import StringHandler
-from app.forms.handlers.text_area_handler import TextAreaHandler
+from app.forms.field_handlers.date_handler import DateHandler
+from app.forms.field_handlers.dropdown_handler import DropdownHandler
+from app.forms.field_handlers.duration_handler import DurationHandler
+from app.forms.field_handlers.number_handler import NumberHandler
+from app.forms.field_handlers.select_handler import SelectHandler
+from app.forms.field_handlers.select_multiple_handler import SelectMultipleHandler
+from app.forms.field_handlers.string_handler import StringHandler
+from app.forms.field_handlers.text_area_handler import TextAreaHandler
 
 FIELD_HANDLER_MAPPINGS = {
     'Checkbox': SelectMultipleHandler,
