+++ app/forms/questionnaire_form.py
@@ -8,8 +8,8 @@ from flask_wtf import FlaskForm
 from werkzeug.datastructures import MultiDict
 from wtforms import validators
 
-from app.forms.fields import get_field
-from app.forms.handlers.date_handler import DateHandler
+from app.forms.field_factory import get_field
+from app.forms.field_handlers.date_handler import DateHandler
 from app.validation.validators import DateRangeCheck, SumCheck, MutuallyExclusiveCheck
 
 logger = logging.getLogger(__name__)
@@ -221,12 +221,12 @@ class QuestionnaireForm(FlaskForm):
         handler = DateHandler(
             date_from, {}, self.answer_store, self.metadata, location=self.location
         )
-        from_min_period_date, from_max_period_date = handler.get_date_limits()
+        from_min_period_date = handler.get_date_limit('minimum')
+        from_max_period_date = handler.get_date_limit('maximum')
 
-        handler = DateHandler(
-            date_to, {}, self.answer_store, self.metadata, location=self.location
-        )
-        to_min_period_date, to_max_period_date = handler.get_date_limits()
+        handler.answer_schema = date_to
+        to_min_period_date = handler.get_date_limit('minimum')
+        to_max_period_date = handler.get_date_limit('maximum')
 
         min_period_date = from_min_period_date or from_max_period_date
         max_period_date = to_max_period_date or to_min_period_date
