+++ tests/app/forms/handlers/test_field_handler.py
@@ -1,7 +1,7 @@
 from wtforms import validators
 
 from app.data_model.answer_store import AnswerStore
-from app.forms.handlers.string_handler import StringHandler
+from app.forms.field_handlers.string_handler import StringHandler
 from app.validation.validators import ResponseRequired
 
 
@@ -9,7 +9,7 @@ def test_get_mandatory_validator_optional():
     answer = {'mandatory': False}
 
     text_area_handler = StringHandler(answer, None, AnswerStore(), {})
-    validate_with = text_area_handler.get_mandatory_validator('MANDATORY_TEXTFIELD')
+    validate_with = text_area_handler.get_mandatory_validator()
 
     assert isinstance(validate_with[0], validators.Optional)
 
@@ -24,7 +24,7 @@ def test_get_mandatory_validator_mandatory():
         {},
     )
 
-    validate_with = text_area_handler.get_mandatory_validator('MANDATORY_TEXTFIELD')
+    validate_with = text_area_handler.get_mandatory_validator()
 
     assert isinstance(validate_with[0], ResponseRequired)
     assert validate_with[0].message == 'This is the default mandatory message'
@@ -46,7 +46,7 @@ def test_get_mandatory_validator_mandatory_with_error():
         AnswerStore(),
         {},
     )
-    validate_with = text_area_handler.get_mandatory_validator('MANDATORY_TEXTFIELD')
+    validate_with = text_area_handler.get_mandatory_validator()
 
     assert isinstance(validate_with[0], ResponseRequired)
     assert validate_with[0].message == 'This is the mandatory message for an answer'
