+++ tests/app/forms/handlers/test_text_area_handler.py
@@ -1,5 +1,5 @@
 from app.data_model.answer_store import AnswerStore
-from app.forms.handlers.text_area_handler import TextAreaHandler
+from app.forms.field_handlers.text_area_handler import TextAreaHandler
 
 
 def test_get_length_validator():
@@ -9,11 +9,9 @@ def test_get_length_validator():
         AnswerStore(),
         {},
     )
-    validate_with = text_area_handler.get_length_validator()
+    validator = text_area_handler.get_length_validator()
 
-    assert (
-        validate_with[0].message == 'This is the default max length of %(max)d message'
-    )
+    assert validator.message == 'This is the default max length of %(max)d message'
 
 
 def test_get_length_validator_with_message_override():
@@ -31,9 +29,9 @@ def test_get_length_validator_with_message_override():
         {},
     )
 
-    validate_with = text_area_handler.get_length_validator()
+    validator = text_area_handler.get_length_validator()
 
-    assert validate_with[0].message == 'A message with characters %(max)d placeholder'
+    assert validator.message == 'A message with characters %(max)d placeholder'
 
 
 def test_get_length_validator_with_max_length_override():
@@ -42,6 +40,6 @@ def test_get_length_validator_with_max_length_override():
     text_area_handler = TextAreaHandler(
         answer, {'MAX_LENGTH_EXCEEDED': '%(max)d characters'}, AnswerStore(), {}
     )
-    validate_with = text_area_handler.get_length_validator()
+    validator = text_area_handler.get_length_validator()
 
-    assert validate_with[0].max == 30
+    assert validator.max == 30
